feat: split app component to children components
