--- src/components/Button.tsx
@@ -1,5 +1,0 @@
-// import React from 'react';
-
-// export const Button = (props) => {
-//   return <button className="page-btn">{props.number}</button>;
-// };
