+++ src/components/Buttons/ButtonTypes.ts
@@ -1,0 +1,5 @@
+export type ButtonProps = {
+  text: string;
+  sendRequest?: () => Promise<void>;
+  makeError?: () => void;
+};
