+++ src/components/ErrorBoundary/ErrorBoundaryTypes.ts
@@ -1,0 +1,11 @@
+import { ErrorInfo, ReactNode } from 'react';
+
+export type ErrorBoundaryProps = {
+  children?: ReactNode;
+};
+
+export type ErrorBoundaryState = {
+  hasError: boolean;
+  error: null | Error;
+  errorInfo: null | ErrorInfo;
+};
