+++ src/components/Pagination/PaginationTypes.ts
@@ -1,0 +1,5 @@
+export type PaginationProps = {
+  totalPages: number;
+  currentPage: number;
+  changePage: (page: number) => void;
+};
