+++ src/components/SearchBar/SearchBarTypes.ts
@@ -1,0 +1,9 @@
+export type SearchBarProps = {
+  query: string;
+  placeholder: string;
+  type: string;
+  className: string;
+  sendRequest?: () => Promise<void>;
+  makeError?: () => void;
+  handleInput?: (event: React.ChangeEvent<HTMLInputElement>) => void;
+};
