+++ src/components/SearchResults/SearchResultsTypes.ts
@@ -1,0 +1,12 @@
+export type SearchResultType = {
+  name: string;
+  height: string;
+  birth: string;
+  race: string;
+  spouse: string;
+};
+
+export type SearchResultProps = {
+  searchResults: SearchResultType[];
+  isSearchLoading: boolean;
+};
