+++ src/main.tsx
@@ -1,10 +1,10 @@
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import './index.css';
-import { ErrorBoundary } from './components/ErrorBoundary';
+import { ErrorBoundary } from './components/ErrorBoundary/ErrorBoundary';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <ErrorBoundary>
-    <App value={''} />
+    <App />
   </ErrorBoundary>
 );
