+++ src/types/types.ts
@@ -1,30 +1,7 @@
-import { ErrorInfo, ReactNode } from 'react';
-
-export type Props = {
-  value: string;
-};
-
-export type SearchResult = {
-  name: string;
-  height: string;
-  birth: string;
-  race: string;
-  spouse: string;
-};
-
+import { SearchResultType } from '../components/SearchResults/SearchResultsTypes';
 export type AppState = {
   query: string;
-  searchResults: SearchResult[];
+  searchResults: SearchResultType[];
   isSearchLoading: boolean;
   hasError: boolean;
-};
-
-export type ErrorBoundaryProps = {
-  children?: ReactNode;
-};
-
-export type ErrorBoundaryState = {
-  hasError: boolean;
-  error: null | Error;
-  errorInfo: null | ErrorInfo;
 };
