+++ src/ulits/pagesFunction.ts
@@ -1,7 +1,3 @@
-// export const getTotalPages = (totalCount: number, limit: number) => {
-//   return Math.ceil(totalCount / limit);
-// };
-
 export const getPagesNumbers = (totalPages: number) => {
   const pagesArr = [];
   for (let i = 0; i < totalPages; i++) {
